refactor: split CPU6502 into modular architecture
- Extract opcodes table to cpu6502/opcodes.ts (1130 lines)
- Extract addressing modes to cpu6502/addressing.ts (90 lines)
- Extract instructions to cpu6502/instructions.ts (645 lines)
- Extract debug functionality to cpu6502/debug.ts (110 lines)
- Main CPU class in cpu6502/core.ts (710 lines)
- Add shared types in cpu6502/types.ts
- Maintain backward compatibility via CPU6502.ts re-export
- All 591 tests passing

This improves maintainability and reduces context window size for AI tools
