+++ src/version.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = '4.33.37';
+export const APP_VERSION = '4.34.0';
